+++ scripts/compress_rdf_vrpt_worker.sh
@@ -63,7 +63,9 @@ do
    mkdir -p $div_dir
   fi
 
-  mv -f $rdf_file $div_dir && gzip $div_dir/$pdb_id-validation-full.rdf
+  if [ ! -s $rdf_file ] ; then
+   mv -f $rdf_file $div_dir && gzip $div_dir/$pdb_id-validation-full.rdf
+  fi
 
  fi
